refactor: 旧 re-export ファイル 4 件削除 + RacingGame.tsx インポートパス更新（5-2）
## 削除した旧ファイル
- drift.ts, wall-physics.ts（参照なし）
- heat.ts, course-effects.ts（index.ts から新モジュールに切替後に削除）

## RacingGame.tsx のインポートパス更新
- Heat → domain/player/heat の updateHeat/getHeatBoost
- CourseEffects → domain/track/course-effect の直接関数
- Highlight.check* → domain/highlight/event-detector の detect*
- computeAllCardEffects → domain/card/card-effect

## highlight.ts の最小化（64行→24行）
- HIGHLIGHT_LABELS/COLORS のみ残し、ドメインロジック re-export を除去

テスト: 27スイート / 245テスト 全パス
型チェック: エラーなし
