+++ src/features/racing-game/index.ts
@@ -16,12 +16,12 @@ export { Logic } from './game-logic';
 export { useInput, useIdle } from './hooks';
 export { VolumeCtrl } from './components/VolumeControl';
 export { DraftCards } from './draft-cards';
-export { Highlight } from './highlight';
+export { HIGHLIGHT_LABELS, HIGHLIGHT_COLORS } from './highlight';
 export { getCardMultiplier, computeAllCardEffects } from './card-effects';
 export type { ComputedCardEffects } from './card-effects';
-export { Heat } from './heat';
-export { CourseEffects } from './course-effects';
-export { calculateWallPenalty, shouldWarp, calculateWarpDestination, calculateSlideVector, calculateSlideAngle, calculateWallSlidePosition } from './wall-physics';
+export { createHeatState, updateHeat, getHeatBoost } from './domain/player/heat';
+export { getCourseEffect, getSegmentFriction, getSegmentSpeedModifier } from './domain/track/course-effect';
+export { calculateWallPenalty, shouldWarp, calculateWarpDestination, calculateSlideVector, calculateSlideAngle, calculateWallSlidePosition } from './domain/track/wall-physics';
 
 // === 新ドメイン層モジュール ===
 export type { GameMode, RaceConfig, RaceState } from './domain/race/types';
