--- src/features/racing-game/wall-physics.ts
@@ -1,11 +1,0 @@
-// 壁衝突物理計算モジュール
-// 移行期間中: domain/track/wall-physics.ts へ委譲
-
-export {
-  calculateWallPenalty,
-  shouldWarp,
-  calculateWarpDestination,
-  calculateSlideVector,
-  calculateSlideAngle,
-  calculateWallSlidePosition,
-} from './domain/track/wall-physics';
